Update .gitignore + requirements.txt.
Limiting NumPy to version < 1.24.0 because np.int deprecation expired
and the method was removed in 1.24.0. Keeping networkx at 2.4 will need
to keep that limitation in place since that's where the error orginated
if using NumPy 1.24.0 and above. The int method was deprecated in
version 1.20.0:
https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations. And
you can see the release notes for version 1.24.0:
https://numpy.org/devdocs/release/1.24.0-notes.html#expired-deprecations.